refactor(frontend/dashboard): move dashboards to role-based feature structure
- Move ParentDashboard.jsx → src/dashboard/parent/components/
- Move TeacherDashboard.jsx → src/dashboard/teacher/components/
- Update service imports to use kebab-case module paths
- Update shared imports (ProfilePhotoModal, useProfilePhoto) to src/shared/*
- Add barrel exports for parent, teacher, and dashboard index
- Add src/dashboard/README.md
- Update App.jsx imports to resolve all moved components
- Update App.test.jsx mock paths to new file locations
- Add argsIgnorePattern to eslint no-unused-vars rule for PascalCase params
