+++ student-board/src/dashboard/index.js
@@ -1,0 +1,2 @@
+export * from './parent/components/index.js';
+export * from './teacher/components/index.js';
