+++ student-board/src/dashboard/parent/components/index.js
@@ -1,0 +1,1 @@
+export { default as ParentDashboard } from './ParentDashboard.jsx';
