+++ student-board/src/dashboard/teacher/components/index.js
@@ -1,0 +1,1 @@
+export { default as TeacherDashboard } from './TeacherDashboard.jsx';
